+++ frontend/src/components/SongTable/DynamicMixModal.tsx
@@ -1,8 +1,9 @@
 import axios from 'axios';
 import * as React from 'react';
-import { Alert, Button, Modal } from 'react-bootstrap';
+import { Button, Modal } from 'react-bootstrap';
 import { DynamicMix } from '../../models/DynamicMix';
 import { SongData } from '../../models/SongData';
+import DynamicMixModalForm from './Form/DynamicMixModalForm';
 
 interface Props {
   song?: SongData;
@@ -14,6 +15,8 @@ interface Props {
 }
 
 interface State {
+  model: string;
+  randomShifts: number;
   errors: string[];
 }
 
@@ -24,15 +27,37 @@ class DynamicMixModal extends React.Component<Props, State> {
   constructor(props: Props) {
     super(props);
     this.state = {
+      model: 'spleeter',
+      randomShifts: 0,
       errors: [],
     };
   }
+
+  /**
+   * Reset all non-error state fields
+   */
+  resetState = (): void => {
+    this.setState({
+      model: 'spleeter',
+      randomShifts: 0,
+      errors: [],
+    });
+  };
 
   /**
    * Called when modal hidden without finishing
    */
   onHide = (): void => {
     this.props.hide();
+    this.props.exit();
+  };
+
+  /**
+   * Called when modal finishes exit animation
+   */
+  onExited = (): void => {
+    // Reset here so modal contents do not flicker during animation
+    this.resetState();
     this.props.exit();
   };
 
@@ -46,8 +71,11 @@ class DynamicMixModal extends React.Component<Props, State> {
 
     const data = {
       source_track: this.props.song.id,
+      separator: this.state.model,
+      random_shifts: this.state.randomShifts,
       overwrite: true,
     };
+
     // Make API request to create the mix
     axios
       .post<DynamicMix>('/api/mix/dynamic/', data)
@@ -63,6 +91,18 @@ class DynamicMixModal extends React.Component<Props, State> {
       });
   };
 
+  handleModelChange = (event: React.ChangeEvent<HTMLSelectElement>): void => {
+    const { value } = event.target;
+    this.setState({ model: value });
+    console.log('model change:', value);
+  };
+
+  handleRandomShiftsChange = (event: React.ChangeEvent<HTMLSelectElement>): void => {
+    const { value } = event.target;
+    this.setState({ randomShifts: parseInt(value) });
+    console.log('rand shift change:', parseInt(value));
+  };
+
   render(): JSX.Element | null {
     const { errors } = this.state;
     const { show, song } = this.props;
@@ -71,19 +111,17 @@ class DynamicMixModal extends React.Component<Props, State> {
     }
 
     return (
-      <Modal show={show} onHide={this.onHide}>
+      <Modal show={show} onHide={this.onHide} onExited={this.onExited}>
         <Modal.Header closeButton>
           <Modal.Title>Create dynamic mix</Modal.Title>
         </Modal.Header>
         <Modal.Body>
-          Are you sure you want to create the mix? This will take a couple of minutes.
-          {errors.length > 0 && (
-            <Alert variant="danger">
-              {errors.map((val, idx) => (
-                <div key={idx}>{val}</div>
-              ))}
-            </Alert>
-          )}
+          <DynamicMixModalForm
+            song={song}
+            errors={errors}
+            handleModelChange={this.handleModelChange}
+            handleRandomShiftsChange={this.handleRandomShiftsChange}
+          />
         </Modal.Body>
         <Modal.Footer>
           <Button variant="outline-danger" onClick={this.onHide}>
