+++ frontend/src/components/SongTable/PausePlayButton.tsx
@@ -24,20 +24,8 @@ class PausePlayButton extends React.Component<Props> {
 
   render(): JSX.Element {
     const { playing, disabled } = this.props;
-
-    const customButtonStyle: React.CSSProperties = disabled
-      ? {
-          pointerEvents: 'none',
-        }
-      : {};
     const customButton = (
-      <Button
-        disabled={disabled}
-        style={customButtonStyle}
-        onClick={this.handlePlay}
-        className="btn-circle p-1"
-        variant="secondary"
-        size="lg">
+      <Button disabled={disabled} onClick={this.handlePlay} className="btn-circle p-1" variant="secondary" size="lg">
         {playing && !disabled ? <PauseFill size={28} /> : <PlayFill size={28} />}
       </Button>
     );
